Collate mouse click implementation in unused folder
