--- src/main/java/seedu/taskboss/commons/events/ui/CategoryListPanelViewingChangedEvent.java
@@ -1,26 +1,0 @@
-package seedu.taskboss.commons.events.ui;
-
-import seedu.taskboss.commons.events.BaseEvent;
-import seedu.taskboss.model.category.Category;
-
-//@@author A0143157J
-/**
- * Represents a viewing change in the Category List Panel
- */
-public class CategoryListPanelViewingChangedEvent extends BaseEvent {
-
-    private final Category newViewing;
-
-    public CategoryListPanelViewingChangedEvent(Category newViewing) {
-        this.newViewing = newViewing;
-    }
-
-    @Override
-    public String toString() {
-        return this.getClass().getSimpleName();
-    }
-
-    public Category getNewViewing() {
-        return newViewing;
-    }
-}
