+++ src/main/java/seedu/taskboss/ui/MainWindow.java
@@ -123,14 +123,6 @@ public class MainWindow extends UiPart<Region> {
         categoryListPanel.updateCategoryCountToHashMap();
     }
 
-    //@@author A0143157J
-    /**
-     * Loads tasks from the selected category
-     */
-    public void loadCategorySelection(Category category) {
-        logic.updateFilteredTaskListByCategory(category);
-    }
-
     //@@author
     private AnchorPane getCommandBoxPlaceholder() {
         return commandBoxPlaceholder;
